requirements.txt
anthropic>=0.40.0
youtube-transcript-api>=0.6.0
streamlit>=1.40.0
python-dotenv>=1.0.0
stripe>=10.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
google-auth-oauthlib>=1.2.0
google-auth>=2.29.0
